+++ src/services/__mocks__/data/cart-data.json
@@ -1,0 +1,74 @@
+{
+  "id_new_cart": {
+    "object": "cart",
+    "id": "cart_01G6F3YATX84TZ49T3CRJ1F5W1",
+    "shipping_address": {
+      "id": "addr_01G6F3Z300RV1JA94QC94PKSKB",
+      "created_at": "2022-06-26T04:22:53.633Z",
+      "updated_at": "2022-06-26T04:22:53.633Z",
+      "deleted_at": null,
+      "customer_id": null,
+      "company": null,
+      "first_name": null,
+      "last_name": null,
+      "address_1": null,
+      "address_2": null,
+      "city": null,
+      "country_code": "ca",
+      "province": null,
+      "postal_code": null,
+      "phone": null,
+      "metadata": null
+    },
+    "items": [],
+    "discounts": [],
+    "region": {
+      "id": "reg_01G4XYVBF3SDMZ5MVKBJBN8FTC",
+      "created_at": "2022-06-07T02:10:41.157Z",
+      "updated_at": "2022-06-12T02:38:58.764Z",
+      "deleted_at": null,
+      "name": "Canada",
+      "currency_code": "cad",
+      "tax_rate": 0,
+      "tax_code": null,
+      "gift_cards_taxable": true,
+      "automatic_taxes": true,
+      "tax_provider_id": null,
+      "metadata": null,
+      "payment_providers": [
+        {
+          "id": "manual",
+          "is_installed": true
+        }
+      ],
+      "fulfillment_providers": [
+        {
+          "id": "shippo",
+          "is_installed": true
+        },
+        {
+          "id": "manual",
+          "is_installed": true
+        }
+      ]
+    },
+    "created_at": "2022-06-26T04:22:28.911Z",
+    "updated_at": "2022-06-26T04:22:53.633Z",
+    "deleted_at": null,
+    "email": null,
+    "billing_address_id": null,
+    "shipping_address_id": "addr_01G6F3Z300RV1JA94QC94PKSKB",
+    "region_id": "reg_01G4XYVBF3SDMZ5MVKBJBN8FTC",
+    "customer_id": null,
+    "payment_id": null,
+    "type": "default",
+    "completed_at": null,
+    "payment_authorized_at": null,
+    "idempotency_key": null,
+    "context": {
+      "ip": "",
+      "user_agent": ""
+    },
+    "metadata": null
+  }
+}
