+++ src/services/__mocks__/data/cart.js
@@ -1,0 +1,7 @@
+import { faker } from "@faker-js/faker"
+
+export const mockCart = () => {
+
+  
+
+}
