+++ src/services/__mocks__/data/shippo-data.json
@@ -34,44 +34,6 @@
       "service_levels": []
     }
   ],
-  "service_groups": [
-    {
-      "description": "2 - 8 days",
-      "flat_rate": "15",
-      "flat_rate_currency": "CAD",
-      "free_shipping_threshold_currency": null,
-      "free_shipping_threshold_min": null,
-      "is_active": true,
-      "name": "Expedited Shipping Canada",
-      "object_id": "54drg6c1e4541gdf321v",
-      "rate_adjustment": 0,
-      "service_levels": [
-        {
-          "account_object_id": "5gf1s5fd12fb45das",
-          "service_level_token": "canada_post_expedited_parcel"
-        }
-      ],
-      "type": "LIVE_RATE"
-    },
-    {
-      "description": "4–7 days",
-      "flat_rate": "35",
-      "flat_rate_currency": "USD",
-      "free_shipping_threshold_currency": null,
-      "free_shipping_threshold_min": null,
-      "is_active": true,
-      "name": "Expedited Shipping USA",
-      "object_id": "5be9e4cddbbd42f9a902cda759f72ca8",
-      "rate_adjustment": 0,
-      "service_levels": [
-        {
-          "account_object_id": "5gf1s5fd12fb45das",
-          "service_level_token": "canada_post_expedited_parcel_usa"
-        }
-      ],
-      "type": "LIVE_RATE"
-    }
-  ],
   "service_level_option": [
     {
       "token": "canada_post_tracked_packet_international",
@@ -116,6 +78,130 @@
         "75": "",
         "200": ""
       }
+    }
+  ],
+  "service_groups": [
+    {
+      "description": "2 - 8 days",
+      "flat_rate": "15",
+      "flat_rate_currency": "CAD",
+      "free_shipping_threshold_currency": null,
+      "free_shipping_threshold_min": null,
+      "is_active": true,
+      "name": "Expedited Shipping Canada",
+      "object_id": "54drg6c1e4541gdf321v",
+      "rate_adjustment": 0,
+      "service_levels": [
+        {
+          "account_object_id": "5gf1s5fd12fb45das",
+          "service_level_token": "canada_post_expedited_parcel"
+        }
+      ],
+      "type": "LIVE_RATE"
+    },
+    {
+      "description": "4–7 days",
+      "flat_rate": "35",
+      "flat_rate_currency": "USD",
+      "free_shipping_threshold_currency": null,
+      "free_shipping_threshold_min": null,
+      "is_active": true,
+      "name": "Expedited Shipping USA",
+      "object_id": "dsfg560asef654asr35g",
+      "rate_adjustment": 0,
+      "service_levels": [
+        {
+          "account_object_id": "5gf1s5fd12fb45das",
+          "service_level_token": "canada_post_expedited_parcel_usa"
+        }
+      ],
+      "type": "LIVE_RATE"
+    }
+  ],
+  "fulfillment_options":[
+    {
+      "token": "canada_post_tracked_packet_international",
+      "name": "Canada Post Tracked Packet - International",
+      "supports_return_labels": false,
+      "id": "shippo-fulfillment-canada_post_tracked_packet_international",
+      "carrier_id": "fdg56d15c65g4dfg",
+      "is_group": false,
+      "carrier": "canada_post",
+      "object_id": "fdg56d15c65g4dfg",
+      "object_owner": "",
+      "account_id": "shippo_canada_post_account",
+      "parameters": {},
+      "test": true,
+      "active": true,
+      "is_shippo_account": true,
+      "metadata": "",
+      "carrier_name": "Canada Post",
+      "carrier_images": {
+        "75": "",
+        "200": ""
+      }
+    },
+    {
+      "token": "canada_post_tracked_packet_usa",
+      "name": "Canada Post Tracked Packet - USA",
+      "supports_return_labels": false,
+      "id": "shippo-fulfillment-canada_post_tracked_packet_usa",
+      "carrier_id": "fdg56d15c65g4dfg",
+      "is_group": false,
+      "carrier": "canada_post",
+      "object_id": "fdg56d15c65g4dfg",
+      "object_owner": "",
+      "account_id": "shippo_canada_post_account",
+      "parameters": {},
+      "test": true,
+      "active": true,
+      "is_shippo_account": true,
+      "metadata": "",
+      "carrier_name": "Canada Post",
+      "carrier_images": {
+        "75": "",
+        "200": ""
+      }
+    },
+    {
+      "id": "shippo-fulfillment-54drg6c1e4541gdf321v",
+      "is_group": true,
+      "description": "2 - 8 days",
+      "flat_rate": "15",
+      "flat_rate_currency": "CAD",
+      "free_shipping_threshold_currency": null,
+      "free_shipping_threshold_min": null,
+      "is_active": true,
+      "name": "Expedited Shipping Canada",
+      "object_id": "54drg6c1e4541gdf321v",
+      "rate_adjustment": 0,
+      "service_levels": [
+        {
+          "account_object_id": "5gf1s5fd12fb45das",
+          "service_level_token": "canada_post_expedited_parcel"
+        }
+      ],
+      "type": "LIVE_RATE"
+    },
+    {
+      "id": "shippo-fulfillment-dsfg560asef654asr35g",
+      "is_group": true,
+      "description": "4–7 days",
+      "flat_rate": "35",
+      "flat_rate_currency": "USD",
+      "free_shipping_threshold_currency": null,
+      "free_shipping_threshold_min": null,
+      "is_active": true,
+      "name": "Expedited Shipping USA",
+      "object_id": "dsfg560asef654asr35g",
+      "rate_adjustment": 0,
+      "service_levels": [
+        {
+          "account_object_id": "5gf1s5fd12fb45das",
+          "service_level_token": "canada_post_expedited_parcel_usa"
+        }
+      ],
+      "type": "LIVE_RATE"
     }
   ]
 }
